adds key class
