+++ test/enigma_test.rb
@@ -11,7 +11,6 @@ class EnigmaTest < Minitest::Test
   def test_it_exists
     skip
     assert_instance_of Enigma, @enigma
-    #=> #<Enigma:0x00007ff90f24cb78...>
   end
 
   def test_it_can_convert_a_date_string_to_key
@@ -20,23 +19,10 @@ class EnigmaTest < Minitest::Test
     assert_equal ["1", "0", "2", "5"], actual
   end
 
-  def test_keys_array_can_return_an_array
-    skip
-    actual = @enigma.get_keys_array(["1","2","3","4","5"])
-    assert_equal [12,23,34,45], actual
-  end
-
   def test_it_can_make_date_array
     skip
     expected = [1,0,2,5]
     actual = @enigma.make_date_array("040895")
-    assert_equal expected, actual
-  end
-
-  def test_it_can_make_key_array
-    skip
-    expected = [02,27,71,15]
-    actual = @enigma.make_key_array("02715")
     assert_equal expected, actual
   end
 
@@ -58,7 +44,7 @@ class EnigmaTest < Minitest::Test
   end
 
   def test_it_can_encrypt_a_message_with_key_and_date
-    actual = @enigma.encrypt("hello world", "02715", "040895")
+    actual = @enigma.encrypt("hello world", "02715", "040895", "encrypt")
     expected = {
       encryption: "keder ohulw",
       key: "02715",
